I'e added support for paged calls
Some of the calls are by nature paged, so I'e added support to githubapi.hs along the lines
indicated this morning in the lecture. There are now two calls locally for runClientM in
Guthubapi.hs: a non recursie one and a recursie one. The detail is hidden away from the client usage
in lib.hs by some typing fun. To gie you a hint, the call to a github api method in lib.hs is now a
partial function that gets completed only in runClientPagedM (which adds a page number and env) and
runClientM (which adds the env). So in Lib.hs you will see the calls being more simple, with no env
created. That's paid for in GithubAPI.hs, where the code has got a little more sophisticated.